Controller, authentication, game interaction and data access
The Authentication process is functional and stores serialized accounts in the data package, password checking, logging in, making accounts and checking that usernames are unique is functional. Will need a better way to handle accessing the logged in account that doesn't require constantly serializing in the future. DataAccess works although no one else should really need to use it ever, integrated serializer into DataAccess since the slides specify it has to be on the outer layer, not a use case. The Login and Games controllers are fully functional from the input side, the login presenter calls code is in the authentication runner under login and signupRequest is commented out until Samuel commits. The other controllers will need lots of work.